feat: rebase fix + add more function to shared mock
