+++ test/e2e/tests/hardware-wallets/trezor/mocks.ts
@@ -2,6 +2,7 @@ import { MockttpServer } from 'mockttp';
 import {
   mockEthDaiTrade,
   mockExternalAccountsAPI,
+  mockIcon,
   mockPriceAPIs,
   mockSuggestedGasFees,
   mockSwapAggregatorMetadata,
@@ -16,24 +17,17 @@ import {
 export async function mockTrezorTransactionRequests(mockServer: MockttpServer) {
   await mockTransactionRequestsBase(mockServer);
   await mockEthDaiTrade(mockServer);
+
   await mockSwapNetworkInfo(mockServer);
   await mockSwapFeatureFlags(mockServer);
   await mockSwapTokens(mockServer);
   await mockSwapTopAssets(mockServer);
   await mockSwapAggregatorMetadata(mockServer);
   await mockSwapGasPrices(mockServer);
+
   await mockSuggestedGasFees(mockServer);
   await mockPriceAPIs(mockServer);
-  await mockExternalAccountsAPI(mockServer);
 
-  const tokenIconResponse = {
-    statusCode: 200,
-    body: 'fake-image-data',
-    headers: { 'content-type': 'image/png' },
-  };
-  await mockServer
-    .forGet(
-      /https:\/\/static\.cx\.metamask\.io\/api\/v1\/tokenIcons\/\d+\/0x[a-fA-F0-9]{40}\.png/u,
-    )
-    .thenCallback(() => tokenIconResponse);
+  await mockExternalAccountsAPI(mockServer);
+  await mockIcon(mockServer);
 }
