+++ test/e2e/tests/hardware-wallets/trezor/trezor-swap.spec.ts
@@ -68,6 +68,10 @@ describe('Trezor Swap', function () {
 
         const swapPage = new SwapPage(driver);
         await swapPage.checkPageIsLoaded();
+
+        //  Occasionally, source token is not set automatically in e2e tests for some reason, so we need to do it manually
+        await swapPage.selectSourceToken('TESTETH');
+
         await swapPage.enterSwapAmount('2');
         await swapPage.selectDestinationToken('DAI');
         await swapPage.dismissManualTokenWarning();
@@ -77,6 +81,11 @@ describe('Trezor Swap', function () {
 
         await swapPage.waitForTransactionToComplete();
 
+        await driver.switchToWindowWithTitle(
+          WINDOW_TITLES.ExtensionInFullScreenView,
+        );
+
+        await homePage.checkPageIsLoaded();
         // check activity list
         await homePage.goToActivityList();
